# API_documentation.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Parcheggio API
  description: API per la gestione delle prenotazioni di parcheggio
servers:
  - url: http://localhost:8080/api
paths:
  /{name}/reservations:
    post:
      summary: Crea una nuova prenotazione
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Prenotazione'
      responses:
        '201':
          description: Prenotazione creata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prenotazione'
        '400':
          description: Errore nella richiesta
        '404':
          description: Parcheggio non trovato
        '500':
          description: Errore interno del server
    put:
      summary: Modifica una prenotazione esistente
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: resId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Prenotazione'
      responses:
        '200':
          description: Prenotazione modificata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prenotazione'
        '400':
          description: Errore nella richiesta
        '404':
          description: Prenotazione o parcheggio non trovato
        '500':
          description: Errore interno del server
    delete:
      summary: Cancella una prenotazione
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: resId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Prenotazione cancellata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prenotazione'
        '404':
          description: Prenotazione o parcheggio non trovato
        '500':
          description: Errore interno del server
    get:
      summary: Ottieni tutte le prenotazioni di un parcheggio
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Elenco delle prenotazioni
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prenotazione'
        '404':
          description: Parcheggio non trovato
        '500':
          description: Errore interno del server
components:
  schemas:
    Prenotazione:
      type: object
      properties:
        timeStart:
          type: string
          format: date-time
          description: Ora di inizio della prenotazione
        timeEnd:
          type: string
          format: date-time
          description: Ora di fine della prenotazione
        fee:
          type: number
          description: Tariffa della prenotazione